+++ packages/embedded-provider-core/src/index.ts
@@ -4,4 +4,5 @@ export * from "./embedded-provider";
 export * from "./auth/jwt-auth";
 export * from "./utils/session";
 export * from "./utils/retry";
+export * from "./chains";
 export * from "./constants";